ci: use Tarantool revision in integration workflow
This patch adds the specific Tarantool revision to be checked out in
callee workflows to avoid the inconsistency between the used workflow
and the version of Tarantool that ref we are using. This revision should
be updated during the bump of the workflow to the corresponding
long-term branch.

The patch left the <testing.yml> workflow intact since it uses a
different approach.

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>
(cherry picked from commit c8e49a29f6d148fcea74d4112053b7f20bca3724)