perf: add CMake infrastructure
This commit introduces CMake building scripts for the benches introduced
before. The benchmarks are enabled only if `LUAJIT_ENABLE_PERF` option
is set. For each suite (LuaJIT-benches in this patch set)
`AddBenchTarget()` macro generates 2 targets:
* Target to run all benches and store results in the
  perf/output/<suite_name> directory.
* Target to run all benches via CTest and inspect results in the
  console.

For the LuaJIT-benches there are 2 generated files:
* FASTA_5000000 -- is used as an input for <k-nukleotide.lua> and
                   <revcomp.lua>.
* SUMCOLL_5000.txt -- is used as an input for <sum-file.lua>.

These files and <perf/output> directory are added to the .gitignore files.

Reviewed-by: Sergey Bronnikov <sergeyb@tarantool.org>
Signed-off-by: Sergey Kaplun <skaplun@tarantool.org>